Check RAILS_env

+++ bin/run.sh
@@ -16,8 +16,15 @@
 set -Eeuxo pipefail
 
 if test "${INTEGRATION_TEST_SETUP:-}" = "true" ; then
-  echo "Setting up for integration tests"
-  RAILS_ENV=test bundle exec rake db:reset db:views:schema:load
+  if test "${RAILS_ENV:-}" = "test"; then
+    echo "Setting up for integration tests"
+    RAILS_ENV=test bundle exec rake db:reset db:views:schema:load
+  else
+    echo "You are providing the flag INTEGRATION_TEST_SETUP, but RAILS_ENV"
+    echo "is different from test, which it is a dangerous operation."
+    echo "This script will stop."
+    exit 1
+  fi
 fi
 
 echo "Starting the service"
